Handle self-closing cells
If we get <x:c /> we need to treat that like an blank cell.  Since it doesn't
have a nested <x:t> node, we'll never hit the usual cell assignment branch.  We
also don't get to the "closer" node branch, so we need to do the missing cell
reference dance, too.